Fix handling of current page
We can't push titan into the address of the current page because if we
do, once we create a new buffer, we can't follow the redirect: it
points the old current-page, which now contains the titan scheme.

Also, if we create a new buffer because no elpher buffer is showing
our editor's current page, put that new buffer in elpher mode and set
a current page. If we don't use elpher mode, or doesn't have a current
page, following the redirect runs into errors.