Use `add-hook' instead of `add-to-list'
`add-to-list' causes issues when used in conjunction with another hook.  For
example, in my config I add a hook using the standard `add-hook' function, which
initializes (apparently) `gemini-mode-hook' with /just/ the function I add.
Subsequent calls to `add-hook' convert the singleton to a list and add to it,
but `add-to-list' throws an error because `gemini-mode-hook' isn't a list.

Emacs ... eh?