Load saved pages in elpher-buffer-name
The old solution didn't really work all that well: you look at a page,
then you load the raw copy, you edit it, it finds the buffer
containing the raw copy (great!), reloads, it gets redirected to the
original page (great!). Then you return to the edit buffer, make a
change and save: now it doesn't find the buffer containing the raw
copy (since the Elpher buffer is showing the original page after the
redirect!) and so a new Elpher buffer gets created, which is then
redirected to the original page (now you have two of them). And if you
make another edit, you have three, and so on, forever. This is lame.